+++ src/main/java/com/example/task/dto/AnimalDocument.java
@@ -7,7 +7,7 @@ public class AnimalDocument {
     private String name;
     private String type;
     private String sex;
-    private Integer weight;
-    private Integer cost;
+    private Double weight;
+    private Double cost;
     public Integer category;
 }
